refactor: move icons into separate module
